requirements.txt
pillow
torch
torchvision
numpy
scipy
scikit-image
tqdm
mmdet
mmcv
